On Solaris when mount.mounted is executed for existing nfs mounts this is incorrectly reporting that mount options have changed when they haven't at all. When rw/ro mount option is passed to mount.mounted this code is comparing these values to what mount.active reports.
On solaris for a current ro nfs mount this returns the string 'read-only' in opts
list and for a rw nfs mount this returns the strings 'write' and 'read'. When these
are compared with the mount options supplied to mount.mounted (rw/ro) then it thinks
the mount options have changed and as such it forces a umount/mount of the filesystem.

"Forced unmount and mount because options (rw) changed"

This PR fixes that behaviour so if these particular mount options have not changed
then it will now not force a umount/mount of the nfs mount.